update mocha version
