+++ .gitignore
@@ -5,3 +5,4 @@ coverage
 npm-debug.log
 .idea
 package-lock.json
+.nyc_output
