+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "hof",
   "description": "A bootstrap for HOF projects",
-  "version": "18.1.0",
+  "version": "18.2.0",
   "license": "MIT",
   "main": "index.js",
   "author": "HomeOffice",
@@ -18,7 +18,7 @@
   },
   "scripts": {
     "test": "npm run unit && npm run lint",
-    "unit": "LOG_LEVEL=error istanbul cover _mocha",
+    "unit": "LOG_LEVEL=error nyc _mocha \"test/integration/**/*.spec.js\"",
     "unit:nocov": "LOG_LEVEL=error mocha",
     "lint": "eslint .",
     "ci": "travis-conditions"
@@ -48,17 +48,22 @@
     "chai": "^3.5.0",
     "debug": "^2.2.0",
     "eslint-config-homeoffice": "^2.1.0",
-    "istanbul": "^0.4.3",
-    "mocha": "^2.4.5",
+    "nyc": "^15.1.0",
+    "mocha": "^8.2.1",
     "pre-commit": "^1.1.3",
     "proxyquire": "^1.7.11",
     "request": "^2.79.0",
-    "sinomocha": "^0.2.4",
-    "sinon": "^1.17.5",
-    "sinon-chai": "^2.8.0",
+    "sinon": "^9.2.1",
+    "sinon-chai": "^3.5.0",
     "supertest": "^3.0.0",
-    "supertest-as-promised": "^3.2.0",
     "travis-conditions": "0.0.0"
+  },
+  "mocha": {
+    "reporter": "spec",
+    "require": "test/common.js",
+    "recursive": "true",
+    "timeout": "6000",
+    "exit": "true"
   },
   "optionalDependencies": {
     "hof-theme-govuk": "^5.1.1"
