+++ test/common.js
@@ -6,7 +6,6 @@ global.chai = require('chai').use(require('sinon-chai'));
 global.should = chai.should();
 global.expect = chai.expect;
 global.sinon = require('sinon');
-require('sinomocha')();
 
 process.setMaxListeners(0);
 process.stdout.setMaxListeners(0);
