+++ test/integration/server.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const request = require('supertest-as-promised');
+const request = require('supertest');
 
 const testTag = 'Test-GA-Tag';
 process.env.GA_TAG = testTag;
@@ -8,21 +8,8 @@ process.env.GA_TAG = testTag;
 const bootstrap = require('../../');
 
 const path = require('path');
-const appConfig = {
-  foo: 'bar',
-  bar: 'baz'
-};
 const root = path.resolve(__dirname, '../fixtures');
 
-let behaviourOptions = null;
-let behaviourCalled = false;
-const behaviour = SuperClass => class extends SuperClass {
-  constructor(options) {
-    behaviourCalled = true;
-    behaviourOptions = options;
-    super(options);
-  }
-};
 
 function getHeaders(res, type) {
   let headers = {};
@@ -37,123 +24,25 @@ function getHeaders(res, type) {
   return headers;
 }
 
-describe('bootstrap()', () => {
-
+describe('hof server', () => {
   before(() => {
     bootstrap.configure('root', root);
   });
 
-  beforeEach(() => {
-    behaviourOptions = null;
-    behaviourCalled = false;
-  });
-
-  it('must be given a list of routes', () =>
-    (() => bootstrap()).should.Throw('Must be called with a list of routes')
-  );
-
-  it('routes must each have a set of one or more steps, or one or more pages', () =>
-    (() => bootstrap({
-      routes: [{}]
-    })).should.Throw('Each app must have steps and/or pages')
-  );
-
-  it('a base fields option must be specified if no route fields option is defined', () =>
-    (() => bootstrap({
-      fields: 'fields',
-      routes: [{
-        steps: {}
-      }]
-    })).should.not.Throw()
-  );
-
-  it('a route fields option must be specified if no base fields option is defined', () =>
-    (() => bootstrap({
-      routes: [{
-        fields: 'apps/app_1/fields',
-        steps: {}
-      }]
-    })).should.not.Throw()
-  );
-
-  it('one of base fields or route fields must be specified as an option', () =>
-    (() => bootstrap({
-      routes: [{
-        steps: {}
-      }]
-    })).should.Throw('Set base fields or route fields or both')
-  );
-
-  it('fields option must be valid when specified', () =>
-    (() => bootstrap({
-      fields: 'not_a_valid_path',
-      routes: [{
-        steps: {},
-      }]
-    })).should.Throw(`Cannot find fields at ${root}/not_a_valid_path`)
-  );
-
-  it('route fields option must be valid when specified', () =>
-    (() => bootstrap({
-      routes: [{
-        steps: {},
-        fields: 'not_a_valid_path'
-      }]
-    })).should.Throw(`Cannot find route fields at ${root}/not_a_valid_path`)
-  );
-
-  it('uses defaults when no views option is specified', () =>
-    (() => bootstrap({
-      fields: 'fields',
-      routes: [{
-        steps: {}
-      }]
-    })).should.not.Throw()
-  );
-
-  it('views option must be valid when specified', () =>
-    (() => bootstrap({
-      fields: 'fields',
-      views: 'invalid_path',
-      routes: [{
-        steps: {}
-      }]
-    })).should.Throw(`Cannot find views at ${root}/invalid_path`)
-  );
-
-  it('route views option must be valid when specified', () =>
-    (() => bootstrap({
-      fields: 'fields',
-      routes: [{
-        views: 'invalid_path',
-        steps: {}
-      }]
-    })).should.Throw(`Cannot find route views at ${root}/invalid_path`)
-  );
-
-  it('does not throw if no route views option is specified and the default route views directory does not exist', () =>
-    (() => bootstrap({
-      fields: 'fields',
-      routes: [{
-        name: 'app_3',
-        steps: {}
-      }]
-    })).should.not.Throw()
-  );
-
   describe('with valid routes and steps', () => {
-
-    it('returns the bootstrap interface object', () =>
-      bootstrap({
+    it('returns a 200 for successful deeper health check', () => {
+      const bs = bootstrap({
         fields: 'fields',
         routes: [{
           views: `${root}/apps/app_1/views`,
-          steps: {
-            '/one': {}
-          }
+          steps: {}
         }]
-      }).should.have.all.keys('server', 'stop', 'start', 'use')
-    );
+      });
+      return request(bs.server)
+        .get('/healthz/readiness')
+        .set('Cookie', ['myCookie=1234'])
+        .expect(200);
+    });
 
     it('starts the service and responds successfully', () => {
       const bs = bootstrap({
@@ -486,68 +375,6 @@ describe('bootstrap()', () => {
         .expect(200);
     });
 
-    it('returns a 200 for successful deeper health check', () => {
-      const bs = bootstrap({
-        fields: 'fields',
-        routes: [{
-          views: `${root}/apps/app_1/views`,
-          steps: {}
-        }]
-      });
-      return request(bs.server)
-        .get('/healthz/readiness')
-        .set('Cookie', ['myCookie=1234'])
-        .expect(200);
-    });
-
-    it('can instantiate a custom behaviour for the route', () => {
-      bootstrap({
-        fields: 'fields',
-        routes: [{
-          views: `${root}/apps/app_1/views`,
-          steps: {
-            '/one': {
-              behaviours: behaviour
-            }
-          }
-        }]
-      });
-      behaviourCalled.should.equal(true);
-    });
-
-    it('can pass the app config to controllers', () => {
-      bootstrap({
-        fields: 'fields',
-        appConfig: appConfig,
-        routes: [{
-          views: `${root}/apps/app_1/views`,
-          steps: {
-            '/one': {
-              behaviours: behaviour
-            }
-          }
-        }]
-      });
-      behaviourOptions.appConfig.should.deep.equal(appConfig);
-    });
-
-    it('can pass the confirm step to controllers', () => {
-      bootstrap({
-        fields: 'fields',
-        appConfig: appConfig,
-        routes: [{
-          views: `${root}/apps/app_1/views`,
-          confirmStep: '/summary',
-          steps: {
-            '/one': {
-              behaviours: behaviour
-            }
-          }
-        }]
-      });
-      behaviourOptions.confirmStep.should.equal('/summary');
-    });
-
     it('can extend CSP directives with CSP config', () => {
       const directives = {
         /* eslint-disable quotes */
@@ -796,31 +623,30 @@ describe('bootstrap()', () => {
           pragma.should.eql('no-cache');
         });
     });
-  });
 
-
-  describe('with user defined middleware', () => {
-    it('can mount user defined middleware with `use`', () => {
-      const bs = bootstrap({
-        start: false,
-        fields: 'fields',
-        routes: [{
-          views: `${root}/apps/app_1/views`,
-          steps: {
-            '/one': {}
-          }
-        }]
+    describe('with user defined middleware', () => {
+      it('can mount user defined middleware with `use`', () => {
+        const bs = bootstrap({
+          start: false,
+          fields: 'fields',
+          routes: [{
+            views: `${root}/apps/app_1/views`,
+            steps: {
+              '/one': {}
+            }
+          }]
+        });
+        bs.use((req, res) => {
+          res.json({ respondedFromMiddleware: true });
+        });
+        bs.start();
+        return request(bs.server)
+          .get('/one')
+          .set('Cookie', ['myCookie=1234'])
+          .expect((response) =>
+            response.body.respondedFromMiddleware.should.equal(true)
+          );
       });
-      bs.use((req, res) => {
-        res.json({ respondedFromMiddleware: true });
-      });
-      bs.start();
-      return request(bs.server)
-        .get('/one')
-        .set('Cookie', ['myCookie=1234'])
-        .expect((response) =>
-          response.body.respondedFromMiddleware.should.equal(true)
-        );
     });
   });
 
@@ -1003,13 +829,13 @@ describe('bootstrap()', () => {
           });
       });
 
-     it('set-cookie header is sent', () => {
-       return request(bs.server)
-         .get('/feedback')
-         .expect(200)
-         .expect((res) => {
-           res.headers['set-cookie'].length.should.be.greaterThan(0);
-         });
+      it('set-cookie header is sent', () => {
+        return request(bs.server)
+          .get('/feedback')
+          .expect(200)
+          .expect((res) => {
+            res.headers['set-cookie'].length.should.be.greaterThan(0);
+          });
       });
     });
 
@@ -1031,5 +857,4 @@ describe('bootstrap()', () => {
         });
     });
   });
-
 });
