--- test/mocha.opts
@@ -1,3 +1,0 @@
---require test/common.js
---recursive
---reporter spec
